+++ HISTORY.md
@@ -3,4 +3,4 @@
 ## v0.0.1 / 2017-12-13
 
 * Fork JetWatson/classnames v2.2.5
-* Add bem function wrapper as default export
+* Add prefix function wrapper as default export